Review: Move WebSocketCallback to api.entity.streaming package
